# swagger.yaml
swagger: '2.0'
info:
    description: API for using in AskBid project
    version: 1.0.0
    title: AskBid API
host: localhost:5000
...
schemes:
    - http
    - https
paths:
...
    /user/sign_up:
        post:
            tags:
                - user
            summary: Sign Up
            description: Users want to sign up
            operationId: signUp
            produces:
                - application/json
            parameters:
                - name: state
                  in: body
                  description: Information of a user
                  type: object
                  properties:
                      fullName:
                          type: string
                      email:
                          type: string
                      password:
                          type: string
            responses:
                200:
                    description: successful operation
                    schema:
                        type: object
                        properties:
                            status:
                                type: string
                409:
                    description: The request could not be completed due to a conflict with the current state.
                    schema:
                        type: object
                        properties:
                            status:
                                type: string
    /user/sign_in:
        post:
            tags:
                - user
            summary: Sign In
            description: Users want to sign in
            operationId: signIn
            produces:
                - application/json
            parameters:
                - name: state
                  in: body
                  description: Information of a user
                  type: object
                  properties:
                      email:
                          type: string
                      password:
                          type: string
            responses:
                200:
                    description: successful operation
                    schema:
                        type: object
                        properties:
                            status:
                                type: string
                            token:
                                type: string
                            userID:
                                type: integer
                401:
                    description: Authorization information is missing or invalid.
                    schema:
                        type: object
                        properties:
                            status:
                                type: string
...